// src/generated/models/Stf_organizationalobjectivesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Stf_organizationalobjectivesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Stf_organizationalobjectivesstatecode = keyof typeof Stf_organizationalobjectivesstatecode;
export const Stf_organizationalobjectivesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Stf_organizationalobjectivesstatuscode = keyof typeof Stf_organizationalobjectivesstatuscode;
export const Stf_organizationalobjectivesstf_objectivestatus = {
  1: 'Active',
  2: 'Achieved',
  3: 'Deferred'
} as const;
export type Stf_organizationalobjectivesstf_objectivestatus = keyof typeof Stf_organizationalobjectivesstf_objectivestatus;
export const Stf_organizationalobjectivesstf_objectivetype = {
  1: 'Organizational',
  2: 'Departmental'
} as const;
export type Stf_organizationalobjectivesstf_objectivetype = keyof typeof Stf_organizationalobjectivesstf_objectivetype;

export interface Stf_organizationalobjectivesBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  "pm_OrgOutputKPI@odata.bind"?: string;
  statecode: Stf_organizationalobjectivesstatecode;
  statuscode?: Stf_organizationalobjectivesstatuscode;
  "stf_BU@odata.bind"?: string;
  stf_currentvalue?: number;
  "stf_Department@odata.bind"?: string;
  stf_enddate?: string;
  "stf_Function@odata.bind"?: string;
  stf_objectivedescription?: string;
  stf_objectiveid?: string;
  stf_objectivestatus?: Stf_organizationalobjectivesstf_objectivestatus;
  stf_objectivetype?: Stf_organizationalobjectivesstf_objectivetype;
  stf_organizationalobjectiveid: string;
  "stf_Owner@odata.bind"?: string;
  "stf_ParentTheme@odata.bind"?: string;
  "stf_PrimaryKPI@odata.bind"?: string;
  "stf_Region@odata.bind"?: string;
  stf_startdate?: string;
  stf_targetvalue?: number;
  stf_title?: string;
  stf_year?: string;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Stf_organizationalobjectives extends Stf_organizationalobjectivesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  pm_orgoutputkpiname?: string;
  statecodename?: string;
  statuscodename?: string;
  stf_buname?: string;
  stf_departmentname?: string;
  stf_functionname?: string;
  stf_objectivestatusname?: string;
  stf_objectivetypename?: string;
  stf_ownername?: string;
  stf_owneryominame?: string;
  stf_parentthemename?: string;
  stf_primarykpiname?: string;
  stf_regionname?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  pm_orgoutputkpi?: object;
  _pm_orgoutputkpi_value?: string;
  stf_bu?: object;
  _stf_bu_value?: string;
  stf_department?: object;
  _stf_department_value?: string;
  stf_function?: object;
  _stf_function_value?: string;
  stf_owner?: object;
  _stf_owner_value?: string;
  stf_parenttheme?: object;
  _stf_parenttheme_value?: string;
  stf_primarykpi?: object;
  _stf_primarykpi_value?: string;
  stf_region?: object;
  _stf_region_value?: string;
}